Worked on File New Report.

+++ CrimeStopper/ShareNewReportViewController.m
@@ -10,8 +10,12 @@
 @import QuartzCore;
 #import "ReportSubmittedViewController.h"
 #import "Parse/Parse.h"
+#import "FHSTwitterEngine.h"
 
-@interface ShareNewReportViewController ()
+@interface ShareNewReportViewController () <FHSTwitterEngineAccessTokenDelegate, UITextViewDelegate> {
+    BOOL check;
+    UIActivityIndicatorView *activityIndicator;
+}
 @property (weak, nonatomic) IBOutlet UITextView *tvFacebook;
 @property (weak, nonatomic) IBOutlet UITextView *tvtwitter;
 - (IBAction)skipClicked:(id)sender;
@@ -19,6 +23,7 @@
 @property (weak, nonatomic) IBOutlet UISwitch *switchFacebook;
 @property (weak, nonatomic) IBOutlet UISwitch *switchTwitter;
 @property (strong, nonatomic) NSArray *photoArray;
+@property (weak, nonatomic) IBOutlet UIScrollView *scrollView;
 
 @end
 
@@ -47,6 +52,23 @@
     self.tvtwitter.layer.borderWidth = 0.2f;
     self.tvtwitter.layer.borderColor = [UIColor darkGrayColor].CGColor;
     self.tvtwitter.layer.cornerRadius = 5;
+    
+    [[FHSTwitterEngine sharedEngine]permanentlySetConsumerKey:@"ZVp4v7Y6llHOWBspbTgwatpL9" andSecret:@"MQBdnkXP8hm69auT2ZkzFlVvaukULLJXg4jeXT80b8pn8SHWvT"];
+    [[FHSTwitterEngine sharedEngine]setDelegate:self];
+    [[FHSTwitterEngine sharedEngine]loadAccessToken];
+    
+    check = YES;
+    
+    // initialize activityIndicator and add it to UIToolBar.
+    activityIndicator = [[UIActivityIndicatorView alloc] initWithActivityIndicatorStyle:UIActivityIndicatorViewStyleWhiteLarge];
+    activityIndicator.frame = CGRectMake(0, 0, 40, 40);
+    activityIndicator.center = self.view.center;
+    [self.view addSubview:activityIndicator];
+    
+    UITapGestureRecognizer *singleTap = [[UITapGestureRecognizer alloc] initWithTarget:self action:@selector(resignTextBox)];
+    [self.scrollView addGestureRecognizer:singleTap];
+    
+    //NSLog(@"access token==> %@", [[FHSTwitterEngine sharedEngine] accessToken].description);
 }
 
 - (void)didReceiveMemoryWarning
@@ -56,14 +78,20 @@
 }
 
 - (IBAction)skipClicked:(id)sender {
-    // delete all local images
-    [self deleteAllimageFiles];
-    
-    ReportSubmittedViewController *vc = [[ReportSubmittedViewController alloc] init];
-    [self.navigationController pushViewController:vc animated:YES];
+    [activityIndicator stopAnimating];
+    if (check) {
+        check = NO;
+        // delete all local images
+        [self deleteAllimageFiles];
+        
+        ReportSubmittedViewController *vc = [[ReportSubmittedViewController alloc] init];
+        [self.navigationController pushViewController:vc animated:YES];
+    }
 }
 
 - (IBAction)shareClicked:(id)sender {
+    
+    [activityIndicator startAnimating];
     
     // code for sharing on FB and Twitter
     
@@ -99,10 +127,47 @@
     
     // For Twitter
     if (self.switchTwitter.on) {
-        
+        if ([[NSUserDefaults standardUserDefaults]objectForKey:@"SavedAccessHTTPBody"] == NULL) {
+            // open login page
+            UIViewController *loginController = [[FHSTwitterEngine sharedEngine]loginControllerWithCompletionHandler:^(BOOL success) {
+                [self postOnTwitter];
+            }];
+            [self presentViewController:loginController animated:YES completion:nil];
+        } else {
+            [self postOnTwitter];
+        }
     }
     
     //[self skipClicked:nil];
+}
+
+-(void)postOnTwitter {
+    NSArray *paths = NSSearchPathForDirectoriesInDomains(NSDocumentDirectory, NSUserDomainMask, YES);
+    NSString *documentsDirectoryPath = [paths objectAtIndex:0];
+    NSString *directory = [documentsDirectoryPath stringByAppendingPathComponent:@"fileNewReport/1.png"];
+    
+    [UIApplication sharedApplication].networkActivityIndicatorVisible = YES;
+    dispatch_async(dispatch_get_global_queue(DISPATCH_QUEUE_PRIORITY_DEFAULT, 0), ^{
+        @autoreleasepool {
+            NSString *tweet = self.tvtwitter.text;
+            id returned = [[FHSTwitterEngine sharedEngine] postTweet:tweet withImageData:[NSData dataWithContentsOfFile:directory]];
+            [UIApplication sharedApplication].networkActivityIndicatorVisible = NO;
+            
+            if ([returned isKindOfClass:[NSError class]]) {
+
+            } else {
+                NSLog(@"%@",returned);
+            }
+            
+            dispatch_sync(dispatch_get_main_queue(), ^{
+                @autoreleasepool {
+                    /*UIAlertView *av = [[UIAlertView alloc]initWithTitle:title message:message delegate:nil cancelButtonTitle:@"OK" otherButtonTitles:nil];
+                    [av show];*/
+                    [self skipClicked:nil];
+                }
+            });
+        }
+    });
 }
 
 -(void)postOnFacebook {
@@ -124,11 +189,10 @@
                                  HTTPMethod:@"POST"
                           completionHandler:^(FBRequestConnection *connection, id result, NSError *error) {
                               if (!error) {
-                                  //NSLog(@"result: %@", result);
-                                  //UIAlertView *alert = [[UIAlertView alloc] initWithTitle:nil message:@"Shared on facebook." delegate:self cancelButtonTitle:@"OK" otherButtonTitles:nil, nil];
-                                  //[alert show];
+                                  [self skipClicked:nil];
                               } else {
                                   NSLog(@"%@", error.description);
+                                  [activityIndicator stopAnimating];
                               }
                           }];
 }
@@ -144,6 +208,7 @@
         BOOL success = [fm removeItemAtPath:[NSString stringWithFormat:@"%@/%@", directory, file] error:&error];
         if (!success || error) {
             // it failed.
+            [activityIndicator stopAnimating];
         }
     }
 }
@@ -154,6 +219,35 @@
         [self.tvFacebook resignFirstResponder];
         [self.tvtwitter resignFirstResponder];
     }
+}
+
+- (void)storeAccessToken:(NSString *)accessToken {
+    [[NSUserDefaults standardUserDefaults]setObject:accessToken forKey:@"SavedAccessHTTPBody"];
+}
+
+- (NSString *)loadAccessToken {
+    return [[NSUserDefaults standardUserDefaults]objectForKey:@"SavedAccessHTTPBody"];
+}
+
+-(void)resignTextBox {
+    [self.tvFacebook resignFirstResponder];
+    [self.tvtwitter resignFirstResponder];
+    [self.scrollView setContentOffset:CGPointMake(0, 0) animated:YES];
+}
+
+#pragma mark - UITextView delegate methods
+
+-(BOOL)textViewShouldBeginEditing:(UITextView *)textView {
+    if (textView == self.tvtwitter) {
+        if ([DeviceInfo isIphone5]) {
+            [self.scrollView setContentOffset:CGPointMake(0, 110) animated:YES];
+        } else {
+            [self.scrollView setContentOffset:CGPointMake(0, 195) animated:YES];
+        }
+    } else if (textView == self.tvFacebook) {
+        [self.scrollView setContentOffset:CGPointMake(0, 0) animated:YES];
+    }
+    return YES;
 }
 
 @end
